JNA-related fix suggested by Normen Muller: use getShort instead of getInt on OSX to get file type from native data structure.

+++ svnkit/src/org/tmatesoft/svn/core/internal/util/jna/SVNLinuxUtil.java
@@ -62,7 +62,7 @@ class SVNLinuxUtil {
                     return SVNFileType.NONE;
                 }
                 int mode = SVNFileUtil.isOSX || SVNFileUtil.isBSD ?
-                        ourSharedMemory.getInt(8) : ourSharedMemory.getInt(16);
+                        ourSharedMemory.getShort(8) : ourSharedMemory.getInt(16);
                 int type = mode & 0170000;
                 if (type == 0120000) {
                     return SVNFileType.SYMLINK;
